2 static SVG draws and 1 dynamic graphic
the first two grids are static HTML defined. the third grid is powered by a JavaScript function to draw a 300x300px grid with center lines highlighted. the third grid is a demonstration of how JavaScript can draw dynamic SVG graphics dynacialy wired to an event. [ button onclick=(); ] in this instance...